backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
sentence-transformers
faiss-cpu
numpy
openai
python-dotenv
pillow
